pyproject.toml
[project]
name = "python-asyncio"
version = "0.1.0"
description = "AsyncIO - Asynchronous Programming"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "httpx>=0.28.1",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "scalene>=1.5.54",
]

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]
